Error handling, and type changes

+++ headers/reader.h
@@ -1,9 +1,9 @@
 #ifndef READER_H
 #define READER_H
     typedef struct Reader Reader;
-    Reader* Reader_create(double time_int, int core_quantity);
+    Reader* Reader_create(int time_int, int core_quantity);
     void Reader_destroy(Reader* rd);
-    double Reader_get_time(Reader* rd);
+    int Reader_get_time(Reader* rd);
     int Reader_get_core_quantity(Reader* rd);
     void* reader_thread(void* args);
 #endif
